Begin work on using custom drawing to implement textDecoration
I don't generally like the direction of this.  While Android doesn't seem to have all the capabilities needed to implement textDecoration natively (eg. different color decoration rather than the text color), it would be painful to manually implement.  Width of an underline, for example, is generally a font-weight specific element... who wants to deal with that?